version static

+++ handlers/order_handler.py
@@ -230,18 +230,12 @@ async def food_type(message: Message, state: FSMContext):
                 text=f"{order_text}\n\nПохоже официант изменил меню👀 \n{text}\n\nСтатус заказа: Не готов",
                 reply_markup=get_order_status_keyboard(message.from_user.id)
             )
+            if text:
+                try:
+                    await message.bot.delete_message(-4773383218, msg_id)
+                except:
+                    pass
 
-            # try:
-            await message.bot.delete_message(-4773383218, msg_id)
-            print("AI LAGMAN", foods_lagman)
-            if foods_shashlik != {}:
-                await message.bot.delete_message(-4921594223, msg_id_shashlik)
-            if foods_lagman != {}:
-                print("its work")
-                await message.bot.delete_message(-4973191656, msg_id_lagman)
-
-            # except:
-            #     await message.answer("Сообщение для удаления не найдно")
 
             if foods_shashlik != {}:
                 order_shashlik_text = format_order_text(table_id, foods_shashlik, full_name=f_name)
@@ -249,6 +243,10 @@ async def food_type(message: Message, state: FSMContext):
 
                 text_shashlik = get_diff(foods_shashlik, foods_db_shashlik)
                 if text_shashlik:
+                    try:
+                        await message.bot.delete_message(-4921594223, msg_id_shashlik)
+                    except:
+                        pass
                     msg_shashlik = await message.bot.send_message(
                         -4921594223,
                         text=f"{order_shashlik_text}\n\nПохоже официант изменил меню👀 \n{text_shashlik}\n\nСтатус заказа: Не готов",
@@ -262,8 +260,12 @@ async def food_type(message: Message, state: FSMContext):
                 text_lagman = get_diff(foods_lagman, foods_db_lagman)
 
                 if text_lagman:
+                    try:
+                        await message.bot.delete_message(-4944653262, msg_id_lagman)
+                    except:
+                        pass
                     msg_lagman = await message.bot.send_message(
-                        -4973191656,
+                        -4944653262,
                         text=f"{order_lagman_text}\n\nПохоже официант изменил меню👀 \n{text_lagman}\n\nСтатус заказа: Не готов",
                         reply_markup=get_order_status_keyboard(message.from_user.id)
                     )
@@ -288,7 +290,7 @@ async def food_type(message: Message, state: FSMContext):
                 order_lagman_text = format_order_text(table_id, foods_lagman, full_name=f_name)
 
                 msg_lagman = await message.bot.send_message(
-                    -4973191656,
+                    -4944653262,
                     text=f"{order_lagman_text}\n\nСтатус заказа: Не готов",
                     reply_markup=get_order_status_keyboard(message.from_user.id)
                 )
@@ -404,7 +406,7 @@ async def food_count_input(message: Message, state: FSMContext):
                 "garnish": garnish
             }
         else:
-            foods[food_name]["count"] = count
+            foods[food_name]["count"] += count
 
     except ValueError:
         await message.answer("Введите целое число.")
